Completed OrdersMapper and I find some bugs in the order version.

+++ src/main/java/com/csu/petstorepro/petstore/entity/Lineitem.java
@@ -29,7 +29,7 @@ public class Lineitem implements Serializable {
 
     private Integer quantity;
 
-    private BigDecimal unitprice;
+    private double unitprice;
 
 
 }
